pyproject.toml
[project]
name = "ravenea"
version = "0.1.0"
description = "[ICLR'26] RAVENEA: A Benchmark for Multimodal Retrieval-Augmented Visual Culture Understanding"
readme = "README.md"
requires-python = ">=3.12.9"
dependencies = [
    "accelerate>=1.12.0",
    "bert-score>=0.3.13",
    "bitsandbytes>=0.49.1",
    "datasets>=4.5.0",
    "dotenv>=0.9.9",
    "evaluate>=0.4.6",
    "openai>=2.15.0",
    "pycocoevalcap>=1.2",
    "rich>=14.3.1",
    "scipy>=1.17.0",
    "timm>=1.0.24",
    "torchmetrics>=1.8.2",
    "transformers>=4.57.6",
    "vllm==0.15.1",
    "wandb==0.24.1",
]
